requirements.txt
# 核心依赖
paddlepaddle>=2.4.0
paddleocr>=2.7.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0

# 图像处理
scikit-image>=0.18.0
matplotlib>=3.3.0

# 数据处理
pandas>=1.3.0
pyyaml>=6.0

# GUI框架
tkinter-tooltip>=2.0.0

# 日期处理
python-dateutil>=2.8.0
regex>=2021.0.0

# 配置管理
pydantic>=1.8.0

# 日志
loguru>=0.6.0

# 性能优化
numba>=0.56.0

# 开发依赖
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
pre-commit>=2.15.0

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# 打包发布
setuptools>=60.0.0
wheel>=0.37.0
twine>=3.8.0
